Home page moved from activity into fragment
