--- app/src/main/res/layout/activity_movie_details.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/movie_card_background">
-
-    <ImageButton
-        android:id="@+id/buttonSearch"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/search_button_top_margin"
-        android:layout_marginEnd="@dimen/search_button_end_margin"
-        android:background="@android:color/transparent"
-        android:contentDescription="@string/search_button"
-        android:src="@drawable/ic_search"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/textPopularNow"
-        style="@style/BaseText.PopularNow"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/popular_now_start_margin"
-        android:layout_marginTop="@dimen/popular_now_top_margin"
-        android:text="@string/popular_now"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/buttonSearch" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvCategories"
-        style="@style/BaseRecyclerView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/categories_recycler_view_top_margin"
-        android:orientation="horizontal"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/textPopularNow"
-        tools:listitem="@layout/item_movie_category" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvMovies"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginStart="@dimen/movies_recycler_view_start_margin"
-        android:layout_marginTop="@dimen/movies_recycler_view_top_margin"
-        android:orientation="horizontal"
-        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        app:layout_constraintBottom_toTopOf="@id/ivHomeButton"
-        app:layout_constraintTop_toBottomOf="@id/rvCategories"
-        tools:listitem="@layout/item_movie" />
-
-    <ImageView
-        android:id="@+id/ivHomeButton"
-        style="@style/BottomNavigationButton"
-        android:layout_width="0dp"
-        android:layout_height="@dimen/bottom_navigation_view_height"
-        android:contentDescription="@string/home_button"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintWidth_default="percent"
-        app:layout_constraintWidth_percent=".5"
-        app:srcCompat="@drawable/ic_home" />
-
-    <ImageView
-        android:id="@+id/ivProfileButton"
-        style="@style/BottomNavigationButton"
-        android:layout_width="0dp"
-        android:layout_height="@dimen/bottom_navigation_view_height"
-        android:contentDescription="@string/profile_button"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintWidth_default="percent"
-        app:layout_constraintWidth_percent=".5"
-        app:srcCompat="@drawable/ic_profile" />
-
-    <ImageView
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/active_indicator"
-        android:scaleType="fitXY"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintWidth_default="percent"
-        app:layout_constraintWidth_percent=".5"
-        app:srcCompat="@drawable/ic_active_indicator" />
-</androidx.constraintlayout.widget.ConstraintLayout>
